Remove Multidex declaration
Since the min SDK is 28, it is no longer necessary to use the Multidex library.

See the following for more info: https://developer.android.com/build/multidex#mdex-on-l

-----

This also removes the dependency on Robolectric's `shadows-supportv4` module, since it was removed in Robolectric 4.9.

See the following for more information: https://robolectric.org/using-add-on-modules/#removed-packages